[ao][experimental] moving apot to sub folder
Summary: this is needed to add __init__.py to experimental (which is
needed to allow it to be added to the nightly), these
folders being in the root of experimental make them called by default
and causes a problem because there are matplotlib dependencies

Test Plan:

python test/quantization/core/experimental/test_linear.py
python test/quantization/core/experimental/test_nonuniform_observer.py
python test/quantization/core/experimental/test_quantized_tensor.py
python test/quantization/core/experimental/test_quantizer.py
python test/quantization/core/experimental/test_fake_quantize.py

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
