+++ torch/ao/quantization/experimental/apot/APoT_tensor.py
@@ -1,5 +1,5 @@
 import torch
-from torch.ao.quantization.experimental.quantizer import APoTQuantizer
+from torch.ao.quantization.experimental.apot.quantizer import APoTQuantizer
 
 # class to store APoT quantized tensor
 class TensorAPoT:
