+++ torch/ao/quantization/experimental/apot/observer.py
@@ -7,7 +7,7 @@ import torch
 import itertools
 import matplotlib.pyplot as plt
 from torch.ao.quantization.observer import ObserverBase
-from torch.ao.quantization.experimental.apot_utils import float_to_apot, apot_to_float
+from torch.ao.quantization.experimental.apot.apot_utils import float_to_apot, apot_to_float
 
 # TODO: Consider adding NonUniformQuantizationObserverBase class
 # when more than one non-uniform method is implemented
